refactor(core): split map_controller into map_overlay_* modules (#141)
- map_overlay_types: MapOverlayResult, SpeciesUrlFn, BaseSpeciesFn, VALID_MAP_VIEWS
- map_overlay_theme: inject_map_overlay_theme
- map_overlay_lifer_popups: format_lifer_popup_lines (removed unused helper)
- map_overlay_lifer_map: build_lifer_overlay_map
- map_overlay_visit_map: build_visit_overlay_map (all locations + species filter)
- map_controller: mode normalisation, dispatch, public API unchanged

docs/development.md table notes helper modules; test module docstring clarified.

Made-with: Cursor
